refactor(header): split HeaderActions into separate component files for better modularity
