+++ src/components/Header/Navigation/HeaderActions/LoginButton.component.tsx
@@ -1,0 +1,7 @@
+import { Button } from "../../../UI/Button/Button.component";
+
+export const LoginButton = () => (
+  <Button className="hidden xl:flex text-sm px-5 py-2.5">
+    Zaloguj się
+  </Button>
+);
