+++ src/components/Header/Navigation/NavLinks.component.tsx
@@ -1,8 +1,7 @@
-import { memo } from 'react';
 import { navigationLinks } from "./navigationData";
 import { NavLink } from "./NavLink.component";
 
-export const NavLinks = memo(() => (
+export const NavLinks = () => (
   <nav className="flex items-center gap-10">
     {navigationLinks.map(({ href, label }) => (
       <NavLink key={href} href={href}>
@@ -10,6 +9,4 @@ export const NavLinks = memo(() => (
       </NavLink>
     ))}
   </nav>
-));
-
-NavLinks.displayName = 'NavLinks';
+);
